Refactor code to use Variant model and add FavoritesProvider
This commit refactors the code to use the Variant model instead of the ProductsModel model. It also adds a FavoritesProvider to manage the user's favorite products. The CartProvider has been updated to use the priceAsDouble getter of the Variant model. The FavoriteItem widget has been updated to use the Variant model and the FavoritesProvider. The FavoriteItemListView has been updated to use the items from the FavoritesProvider. The HomeViewContainerDetailsForList widget has been updated to handle adding to favorites.